Move effect system to engine
